[Refactor] Move image/ to internal
This PR is an effort towards reducing the public surface
of APIs. It move image/ to internal/image

(cherry picked from commit 1f5890709fdc17de6f44f42b5138dd0f7e64bc74)
